+++ src/registers/capability.rs
@@ -6,29 +6,30 @@ use core::{convert::TryInto, fmt};
 /// Capability Registers Length
 #[repr(transparent)]
 #[allow(clippy::module_name_repetitions)]
-#[derive(Debug)]
+#[derive(Copy, Clone, Debug)]
 pub struct CapabilityRegistersLength(u8);
 impl CapabilityRegistersLength {
     /// Returns the length of the Capability Registers.
     #[must_use]
-    pub fn get(&self) -> u8 {
+    pub fn get(self) -> u8 {
         self.0
     }
 }
 
 /// Structural Parameters 1
 #[repr(transparent)]
+#[derive(Copy, Clone)]
 pub struct StructuralParameters1(u32);
 impl StructuralParameters1 {
     /// Returns the number of available device slots.
     #[must_use]
-    pub fn number_of_device_slots(&self) -> u8 {
+    pub fn number_of_device_slots(self) -> u8 {
         self.0.get_bits(0..=7).try_into().unwrap()
     }
 
     /// Returns the number of ports.
     #[must_use]
-    pub fn number_of_ports(&self) -> u8 {
+    pub fn number_of_ports(self) -> u8 {
         self.0.get_bits(24..=31).try_into().unwrap()
     }
 }
@@ -43,6 +44,7 @@ impl fmt::Debug for StructuralParameters1 {
 
 /// Structural Parameters 2
 #[repr(transparent)]
+#[derive(Copy, Clone)]
 pub struct StructuralParameters2(u32);
 impl StructuralParameters2 {
     /// Returns the maximum number of the elements the Event Ring Segment Table can contain.
@@ -50,28 +52,28 @@ impl StructuralParameters2 {
     /// Note that the `ERST Max` field of the Structural Parameters 2 register contains the exponential
     /// value, but this method returns the calculated value.
     #[must_use]
-    pub fn event_ring_segment_table_max(&self) -> u16 {
+    pub fn event_ring_segment_table_max(self) -> u16 {
         2_u16.pow(self.erst_max())
     }
 
     /// Returns the number of scratchpads that xHC needs.
     #[must_use]
-    pub fn max_scratchpad_buffers(&self) -> u32 {
+    pub fn max_scratchpad_buffers(self) -> u32 {
         let h = self.max_scratchpad_buffers_hi();
         let l = self.max_scratchpad_buffers_lo();
 
         h << 5 | l
     }
 
-    fn erst_max(&self) -> u32 {
+    fn erst_max(self) -> u32 {
         self.0.get_bits(4..=7)
     }
 
-    fn max_scratchpad_buffers_hi(&self) -> u32 {
+    fn max_scratchpad_buffers_hi(self) -> u32 {
         self.0.get_bits(20..=25)
     }
 
-    fn max_scratchpad_buffers_lo(&self) -> u32 {
+    fn max_scratchpad_buffers_lo(self) -> u32 {
         self.0.get_bits(27..=31)
     }
 }
@@ -89,13 +91,14 @@ impl fmt::Debug for StructuralParameters2 {
 
 /// Capability Parameters 1
 #[repr(transparent)]
+#[derive(Copy, Clone)]
 #[allow(clippy::module_name_repetitions)]
 pub struct CapabilityParameters1(u32);
 impl CapabilityParameters1 {
     /// Returns `true` if the xHC uses 64 byte Context data structures, and `false` if the xHC uses
     /// 32 byte Context data structures.
     #[must_use]
-    pub fn context_size(&self) -> bool {
+    pub fn context_size(self) -> bool {
         self.0.get_bit(2)
     }
 
@@ -103,7 +106,7 @@ impl CapabilityParameters1 {
     /// zero, the list does not exist.
     /// The base address can be calculated by `(MMIO base) + (xECP) << 2`
     #[must_use]
-    pub fn xhci_extended_capabilities_pointer(&self) -> u16 {
+    pub fn xhci_extended_capabilities_pointer(self) -> u16 {
         self.0.get_bits(16..=31).try_into().unwrap()
     }
 }
@@ -118,24 +121,24 @@ impl fmt::Debug for CapabilityParameters1 {
 
 /// Doorbell Offset
 #[repr(transparent)]
-#[derive(Debug)]
+#[derive(Copy, Clone, Debug)]
 pub struct DoorbellOffset(u32);
 impl DoorbellOffset {
     /// Returns the offset of the Doorbell Array from the MMIO base.
     #[must_use]
-    pub fn get(&self) -> u32 {
+    pub fn get(self) -> u32 {
         self.0
     }
 }
 
 /// Runtime Register Space Offset
 #[repr(transparent)]
-#[derive(Debug)]
+#[derive(Copy, Clone, Debug)]
 pub struct RuntimeRegisterSpaceOffset(u32);
 impl RuntimeRegisterSpaceOffset {
     /// Returns the offset of the Runtime Registers from the MMIO base.
     #[must_use]
-    pub fn get(&self) -> u32 {
+    pub fn get(self) -> u32 {
         self.0
     }
 }