+++ src/registers/doorbell.rs
@@ -5,11 +5,12 @@ use core::{convert::TryInto, fmt};
 
 /// The element of the Doorbell Array.
 #[repr(transparent)]
+#[derive(Copy, Clone)]
 pub struct Register(u32);
 impl Register {
     /// Get a doorbell target.
     #[must_use]
-    pub fn doorbell_target(&self) -> u8 {
+    pub fn doorbell_target(self) -> u8 {
         self.0.get_bits(0..=7).try_into().unwrap()
     }