+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
 ## Unreleased - ReleaseDate
+### Added
+- Implement `Copy` and `Clone` for register types.
 
 ## 0.2.4 - 2021-01-22
 ### Added
-- `extended_capabilities::usb_legacy_support_capability::UsbLegacySupportCapability`
+- `extended_capabilities::usb_legacy_support_capability::UsbLegacySupportCapability`.
 
 ## 0.2.3 - 2021-01-22
 ### Fixed
@@ -12,7 +14,7 @@
 
 ## 0.2.2 - 2021-01-22
 ### Added
-- Implement `Debug` for `error::Error`
+- Implement `Debug` for `error::Error`.
 
 ## 0.2.1 - 2021-01-22
 ### Added
@@ -20,7 +22,7 @@
 
 ## 0.2.0 - 2021-01-22
 ### Changed
-- Rename `EventRingDequeuePointerRegister::set` to `EventRingDequeuePointerRegister::set_event_ring_dequeue_pointer`
+- Rename `EventRingDequeuePointerRegister::set` to `EventRingDequeuePointerRegister::set_event_ring_dequeue_pointer`.
 
 ## 0.1.0 - 2021-01-22
 ### Added